create generic pipeline class
