+++ source/pipelines/GraphicsPipeline.cpp
@@ -1,16 +1,10 @@
 #include "GraphicsPipeline.h"
 #include <stdexcept>
 
-GraphicsPipeline::GraphicsPipeline(std::shared_ptr<PhysicalDevice> physicalDevice,
-                                   std::shared_ptr<LogicalDevice> logicalDevice)
-  : physicalDevice(std::move(physicalDevice)), logicalDevice(std::move(logicalDevice))
+GraphicsPipeline::GraphicsPipeline(const std::shared_ptr<PhysicalDevice> &physicalDevice,
+                                   const std::shared_ptr<LogicalDevice> &logicalDevice)
+  : Pipeline(physicalDevice, logicalDevice)
 {}
-
-GraphicsPipeline::~GraphicsPipeline()
-{
-  vkDestroyPipeline(logicalDevice->getDevice(), pipeline, nullptr);
-  vkDestroyPipelineLayout(logicalDevice->getDevice(), pipelineLayout, nullptr);
-}
 
 void GraphicsPipeline::createShader(const char *filename, VkShaderStageFlagBits stage)
 {
